move updateState to the interface Context
