+++ dev/conductor/core/lib/src/context.dart
@@ -1,4 +1,9 @@
+import 'package:file/file.dart' show File;
+
 import 'globals.dart';
+import 'proto/conductor_state.pb.dart' as pb;
+import 'repository.dart';
+import 'state.dart';
 import 'stdio.dart' show Stdio;
 
 /// Interface for shared functionality across all sub-commands.
@@ -6,7 +11,13 @@ import 'stdio.dart' show Stdio;
 /// Different frontends (e.g. CLI vs desktop) can share [Context]s, although
 /// methods for capturing user interaction may be overridden.
 abstract class Context {
-  const Context();
+  const Context({
+    required this.checkouts,
+    required this.stateFile,
+  });
+
+  final Checkouts checkouts;
+  final File stateFile;
 
   /// Confirm an action with the user before proceeding.
   ///
@@ -25,5 +36,14 @@ abstract class Context {
     throw ConductorException(
       'Unknown user input (expected "y" or "n"): $response',
     );
+  }
+
+  /// Save the release's [state].
+  ///
+  /// This can be overridden by frontends that may not persist the state to
+  /// disk, and/or may need to call additional update hooks each time the state
+  /// is updated.
+  void updateState(pb.ConductorState state, [List<String> logs = const <String>[]]) {
+    writeStateToFile(stateFile, state, logs);
   }
 }
